Reorganise approach 1.
