+++ pdr_backend/predictoor/approach1/main1.py
@@ -1,12 +1,11 @@
 from enforce_typing import enforce_types
 
 from pdr_backend.predictoor.approach1.predictoor_agent1 import PredictoorAgent1
-from pdr_backend.predictoor.approach1.predictoor_config1 import PredictoorConfig1
 
 
 @enforce_types
 def do_main1():
-    config = PredictoorConfig1()
+    config = PredictoorAgent1.predictoor_config_class()
     p = PredictoorAgent1(config)
     p.run()
