+++ pdr_backend/predictoor/base_predictoor_config.py
@@ -24,11 +24,16 @@ to predict in every block during the last 60 seconds before the epoch
 concludes.
 """
 
+from os import getenv
+
 from enforce_typing import enforce_types
 
-from pdr_backend.predictoor.base_predictoor_config import BasePredictoorConfig
+from pdr_backend.models.base_config import BaseConfig
 
 
 @enforce_types
-class PredictoorConfig1(BasePredictoorConfig):
-    pass
+class BasePredictoorConfig(BaseConfig):
+    def __init__(self):
+        super().__init__()
+        self.s_until_epoch_end = int(getenv("SECONDS_TILL_EPOCH_END", "60"))
+        self.stake_amount = int(getenv("STAKE_AMOUNT", "1"))
